move interpolation code
