+++ src/openms/include/OpenMS/ML/INTERPOLATION/sources.cmake
@@ -1,13 +1,10 @@
 ### the directory name
-set(directory include/OpenMS/MATH/MISC)
+set(directory include/OpenMS/ML/INTERPOLATION)
 
 ### list all header files of the directory here
 set(sources_list_h
-BSpline2d.h
-CubicSpline2d.h
-EmgGradientDescent.h
-MathFunctions.h
-SplineBisection.h
+BilinearInterpolation.h
+LinearInterpolation.h
 )
 
 ### add path to the filenames
@@ -17,6 +14,7 @@ foreach(i ${sources_list_h})
 endforeach(i)
 
 ### source group definition
-source_group("Header Files\\OpenMS\\MATH\\MISC" FILES ${sources_h})
+source_group("Header Files\\OpenMS\\ML\\INTERPOLATION" FILES ${sources_h})
 
 set(OpenMS_sources_h ${OpenMS_sources_h} ${sources_h})
+
