+++ src/openms/includes.cmake
@@ -11,6 +11,7 @@ include(source/ML/NNLS/sources.cmake)
 include(source/ML/SVM/sources.cmake)
 include(source/ML/CLUSTERING/sources.cmake)
 include(source/ML/GRIDSEARCH/sources.cmake)
+include(source/ML/INTERPOLATION/sources.cmake)
 include(source/ML/ROCCURVE/sources.cmake)
 include(source/ML/RANSAC/sources.cmake)
 include(source/ML/REGRESSION/sources.cmake)
@@ -62,6 +63,7 @@ include(include/OpenMS/INTERFACES/sources.cmake)
 include(include/OpenMS/CONCEPT/sources.cmake)
 include(include/OpenMS/SYSTEM/sources.cmake)
 include(include/OpenMS/MATH/MISC/sources.cmake)
+include(include/OpenMS/ML/INTERPOLATION/sources.cmake)
 include(include/OpenMS/ML/NNLS/sources.cmake)
 include(include/OpenMS/ML/SVM/sources.cmake)
 include(include/OpenMS/ML/RANSAC/sources.cmake)
