+++ src/openms/source/ML/INTERPOLATION/LinearInterpolation.cpp
@@ -7,7 +7,7 @@
 // --------------------------------------------------------------------------
 //
 
-#include <OpenMS/MATH/MISC/LinearInterpolation.h>
+#include <OpenMS/ML/INTERPOLATION/LinearInterpolation.h>
 
 namespace OpenMS
 {
