+++ src/tests/class_tests/openms/source/BilinearInterpolation_test.cpp
@@ -9,7 +9,7 @@
 ///////////////////////////
 
 // This one is going to be tested.
-#include <OpenMS/MATH/MISC/BilinearInterpolation.h>
+#include <OpenMS/ML/INTERPOLATION/BilinearInterpolation.h>
 
 ///////////////////////////
